+++ src/components/sounds/MyPlayLists.jsx
@@ -1,7 +1,5 @@
 import React, { useEffect, useState } from "react";
 import { useCookies } from "react-cookie";
-import { useNavigate } from "react-router-dom";
-import noResults from "../../assets/images/no-results.svg";
 import { useTranslation } from "react-i18next";
 import axios from "./../../util/axios";
 import DataLoader from "../ui/DataLoader";
@@ -10,10 +8,12 @@ import lib from "../../assets/images/lib.svg";
 import ListCard from "./ListCard";
 import { toast } from "react-toastify";
 import ConfirmDeleteModal from "../ui/ConfirmDeleteModal";
+import noResults from "../../assets/images/no-results.svg";
+import { Link } from "react-router-dom";
+import loginImage from "../../assets/images/login.webp";
 
 const MyPlayLists = () => {
   const { t } = useTranslation();
-  const navigate = useNavigate();
   const [cookies] = useCookies(["refreshToken"]);
   const isAuthenticated = cookies.refreshToken ? true : false;
   const [showModal, setShowModal] = useState(false);
@@ -22,11 +22,6 @@ const MyPlayLists = () => {
   const [playLists, setPlayLists] = useState([]);
   const [playList, setPlayList] = useState({});
   const [loading, setLoading] = useState(false);
-  useEffect(() => {
-    if (!isAuthenticated) {
-      navigate("/login");
-    }
-  }, [isAuthenticated, navigate]);
 
   useEffect(() => {
     const fetchLibrary = async () => {
@@ -37,7 +32,7 @@ const MyPlayLists = () => {
           setPlayLists(libraryResponse?.data?.message);
         }
       } catch (error) {
-        console.log(error);
+        console.error(error);
       } finally {
         setLoading(false);
       }
@@ -47,6 +42,8 @@ const MyPlayLists = () => {
   }, []);
 
   const deleteList = async () => {
+    if (!targetIdForRemove) return;
+
     try {
       const res = await axios.delete(
         `/members/Delete_audio_list/${targetIdForRemove}/`
@@ -57,11 +54,11 @@ const MyPlayLists = () => {
         setShowDeleteModal(false);
       }
     } catch (error) {
-      console.log(error);
+      console.error(error);
     }
   };
 
-  const editList = async (playList) => {
+  const editList = (playList) => {
     setShowModal(true);
     setPlayList(playList);
   };
@@ -77,38 +74,56 @@ const MyPlayLists = () => {
         <DataLoader minHeight="300px" />
       ) : (
         <>
-          {playLists.length === 0 ? (
-            <div className="col-12 p-2">
-              <div className="noDataFound">
-                <img src={noResults} alt="no results" />
-                <h5>{t("sounds.noPlayLists")}</h5>
-                <p>{t("sounds.noPlayListsSub")}</p>
-                <button onClick={() => setShowModal(true)}>
-                  {t("sounds.createPlayList")}
-                </button>
-              </div>
-            </div>
-          ) : (
+          {isAuthenticated ? (
             <>
-              <div className="col-12 p-2 mb-3">
-                <div className="swiper_pagination_title">
-                  <h5>
-                    <img src={lib} alt="list" /> {t("sounds.myLists")}
-                  </h5>
-                  <div className="create_list">
+              {playLists.length === 0 ? (
+                <div className="col-12 p-2">
+                  <div className="noDataFound">
+                    <img src={noResults} alt="no results" />
+                    <h5>{t("sounds.noPlayLists")}</h5>
+                    <p>{t("sounds.noPlayListsSub")}</p>
                     <button onClick={() => setShowModal(true)}>
-                      <i className="fa-light fa-plus"></i>{" "}
                       {t("sounds.createPlayList")}
                     </button>
                   </div>
                 </div>
-              </div>
-              {playLists?.map((list) => (
-                <div className="col-lg-4 p-2" key={list?.id}>
-                  <ListCard list={list} onDelete={getId} onEdit={editList} />
-                </div>
-              ))}
+              ) : (
+                <>
+                  <div className="col-12 p-2 mb-3">
+                    <div className="swiper_pagination_title">
+                      <h5>
+                        <img src={lib} alt="list" /> {t("sounds.myLists")}
+                      </h5>
+                      <div className="create_list">
+                        <button onClick={() => setShowModal(true)}>
+                          <i className="fa-light fa-plus"></i>{" "}
+                          {t("sounds.createPlayList")}
+                        </button>
+                      </div>
+                    </div>
+                  </div>
+                  {playLists?.map((list) => (
+                    <div className="col-lg-4 p-2" key={list?.id}>
+                      <ListCard
+                        list={list}
+                        onDelete={getId}
+                        onEdit={editList}
+                      />
+                    </div>
+                  ))}
+                </>
+              )}
             </>
+          ) : (
+            <div className="col-12 p-2">
+              <div className="noDataFound">
+                <img src={loginImage} alt="no results" className="mb-3" />
+                <h5>{t("sounds.noPlayListsLoginMessage")}</h5>
+                <Link to="/login" className="mt-2">
+                  {t("login")}
+                </Link>
+              </div>
+            </div>
           )}
         </>
       )}
